// src/interfaces/formulario.interface.ts
export interface Formulario {
    nombre?: string;
    paterno?: string;
    materno?: string;
    nacimiento?: string;
    curp?: string;
    sexo?: string;
    estado_civil?: string;
    calle?: string;
    num_interior?: number;
    colonia?: string;
    cp?: number;
    ciudad?: string;
    municipio?: string;
    estado?: string;
    tel?: number;
    matricula?: string;
    celular?: number;
    tel_paterno?: number;
    facebook?: string;
    twitter?: string;
    email?: string;
    carrera?: string;
    especialidad?: string;
    egreso?: string;
    titulado?: string;
    ingles?: number;
    idioma_porcentaje?: number;
    idioma_nombre?: string;
    softwares?: string;
    docentes?: number;
    plan_estudios?: number;
    investigacion_desarrollo?: number;
    investigacion_enseñanza?: number;
    infraestructura?: number;
    residencia?: number;
    iii1?: number;
    iii1_esp_select?: number;
    iii1_esp_input?: string;
    iii1_otra?: string;
    iii2?: number;
    iii3?: number;
    iii4?: number;
    iii4_o?: number;
    iii5?: number;
    iii5_o?: string;
    iii6_hablar?: number;
    iii6_escribir?: number;
    iii6_leer?: number;
    iii6_escuchar?: number;
    iii6?: number;
    iii7?: number;
    iii7_ingreso?: number;
    iii8?: number;
    iii9?: number;
    iii10?: number;
    iii10_o?: number;
    iii11?: number;
    iii12?: number;
    iii12_giro?: string;
    iii12_razon?: string;
    iii12_calle?: string;
    iii12_numero?: number;
    iii12_colonia?: string;
    iii12_cp?: number;
    iii12_ciudad?: string;
    iii12_municipio?: string;
    iii12_estado?: string;
    iii12_tel?: number;
    iii12_fax?: number;
    iii12_email?: string;
    iii12_web?: string;
    iii12_nombre?: string;
    iii12_puesto?: string;
    iii13_primario?: number;
    iii13_secundario?: number;
    iii13_terciario?: number;
    iii14?: number;
    iv1?: number;
    iv2?: number;
    iv3?: number;
    iv4_1?: number;
    iv4_2?: number;
    iv4_3?: number;
    iv4_4?: number;
    iv4_5?: number;
    iv4_6?: number;
    iv4_7?: number;
    iv4_8?: number;
    iv4_9?: number;
    iv4_10?: number;
    v1_a?: number;
    v1_b?: string;
    v2_a?: number;
    v2_b?: string;
    vi1?: number;
    vi1_b?: string;
    vi2?: number;
    vi2_b?: string;
    vi3?: number;
    vii?: string;

}